Add a button to allow hiding the right rail
This uses the checkbox hack, making it usable without JavaScript.
JavaScript enhancements are supported, and to make this possible,
the scripts module now uses packageFiles.

The button is connected to the mirage-show-right-rail preference,
which can be manually toggled in Special:Preferences. This allows
it to be used even when JavaScript is disabled or not available.

Apparently, CSSJanus does not like ~. A workaround with manual
flipping was used instead. This requires not using Less nesting,
as that mangles the /* @noflip */ comments.
Upstream: https://github.com/cssjanus/cssjanus/issues/85.

Change-Id: I7b2109ec0a2fd14576f06bfb7fd6e637730efff5