Use interface-message-box feature
This removes the copy of the style file of that feature in legacy.less.
One change has been made: The border-radius has been removed, as these
message boxes now adhere to the Wikimedia Design Style Guide. This
removes the remaining difference between the OOUI implementation and
the native MediaWiki implementation in Mirage.

Depends-On: I86242c7fa32a32d5f98a2493abf8990412606b2e
Change-Id: I0b4d749461606d6837a5d1be3306310a26b2e0eb