resourceloader: Fix wordmark and tagline from preventing CSS to load
The code assumes that $logo['wordmark'] and $logo['tagline'] are strings, like
$logo['icon'] and $logo['svg']. However, the first two are actually arrays.

This commit fixes getting the wordmark and tagline URLs, preventing loading
skins.mirage.styles from causing an exception.

This commit only preloads $logo['wordmark']['src'] and not the 1x version
because the code only uses the src version. The code should be changed to use
both the 1x and src versions when applicable, though it's not this commit's
place to do so.

Bug: T367381
Change-Id: Ia40e7c9aecac3400bf6fa91649a80d763be9a377